DSR/requirements.txt
asgiref==3.7.2
Django==5.0.1
django-extensions==3.2.3
django-modeladmin-reorder==0.3.1
djangorestframework==3.14.0
pillow==10.2.0
psycopg==3.1.17
python-dotenv==1.0.1
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.9.0
ulid-py==1.1.0
